requirements.txt
pandas
openpyxl
reportlab
python-dotenv
openai